🔀 [BE] merge: from feature/BE-exercise-create/127 to dev

See merge request s12-final/S12P31S206!11

+++ BE/health-service/src/main/java/com/s206/health/exercise/entity/ExerciseCategory.java
@@ -19,7 +19,7 @@ public class ExerciseCategory {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "exercise_category_id", nullable = false)
-    private Long exerciseCategoryId;
+    private Integer exerciseCategoryId;
 
     @Column(name = "exercise_category_name", nullable = false, length = 50)
     private String exerciseCategoryName;
