+++ BE/health-service/src/main/java/com/s206/health/exercise/entity/ExerciseType.java
@@ -19,10 +19,10 @@ public class ExerciseType {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "exercise_type_id")
-    private Long exerciseTypeId;
+    private Integer exerciseTypeId;
 
     @Column(name = "exercise_category_id", nullable = false)
-    private Long exerciseCategoryId;
+    private Integer exerciseCategoryId;
 
     @Column(name = "exercise_name", nullable = false, length = 50)
     private String exerciseName;
